Module Instances.Writer

Parameters

module M : MONOID

Signature

val return : 'a -> 'a * M.t

return x should be a monad which produces x

val map : ( 'a1 -> 'a2 ) -> ('a1 * M.t) -> 'a2 * M.t

Run a monad and apply a function to the result

val bind : ('a1 * M.t) -> ( 'a1 -> 'a2 * M.t ) -> 'a2 * M.t

Sequential composition: First run one monad, then another based on the first result

val let+ : ('a1 * M.t) -> ( 'a1 -> 'a2 ) -> 'a2 * M.t

This is just syntactic sugar for map: let+ x = m in f x is the same as map f m.

val let* : ('a1 * M.t) -> ( 'a1 -> 'a2 * M.t ) -> 'a2 * M.t

This is just syntactic sugar for bind: let* x = m in f x is the same as bind m f.

val (>>=) : ('a1 * M.t) -> ( 'a1 -> 'a2 * M.t ) -> 'a2 * M.t

Operator version of bind, i.e. m1 >>= m2 is the same as bind m1 m2

val (>>) : ('a1 * M.t) -> ('a2 * M.t) -> 'a2 * M.t

Like (>>=) but ignores the result of the first monad, i.e. m1 >> m2 is the same as m1 >>= fun _ -> m2

val (|>>) : ('a * M.t) -> ( 'a -> 'b ) -> 'b * M.t

This is the operator version of map, i.e. m |>> f is the same as map f m

val foldM : ( 'a2 -> 'a1 -> 'a2 * M.t ) -> 'a2 -> 'a1 list -> 'a2 * M.t

Monadic version of List.fold_left: the folding function computes the new accumulator value inside the monad.

val fold1M : ( 'a -> 'a -> 'a * M.t ) -> 'a list -> 'a * M.t

Same as foldM but the first element of the list is taken as the initial accumulator value. Throws an Invalid_argument exception on the empty list.

val mapM : ( 'a -> 'b * M.t ) -> 'a list -> 'b list * M.t

Monadic version of List.map: the new elements are computed inside the monad.

val forM : 'a list -> ( 'a -> 'b * M.t ) -> 'b list * M.t

This is just mapM with its arguments flipped.

val iterM : ( 'a -> unit * M.t ) -> 'a list -> unit * M.t

Monadic version of List.iter

val ifM : bool -> ( unit -> unit * M.t ) -> unit * M.t

Conditionally run the given monad or just return ()

val write : M.t -> unit * M.t