Module type Mnd.MONAD_WITH_AND

include MONAD

include MONAD_DEF

type 'a t

val return : 'a -> 'a t

return x should be a monad which produces x

val map : ( 'a1 -> 'a2 ) -> 'a1 t -> 'a2 t

Run a monad and apply a function to the result

val bind : 'a1 t -> ( 'a1 -> 'a2 t ) -> 'a2 t

Sequential composition: First run one monad, then another based on the first result

val let+ : 'a1 t -> ( 'a1 -> 'a2 ) -> 'a2 t

This is just syntactic sugar for map: let+ x = m in f x is the same as map f m.

val let* : 'a1 t -> ( 'a1 -> 'a2 t ) -> 'a2 t

This is just syntactic sugar for bind: let* x = m in f x is the same as bind m f.

val (>>=) : 'a1 t -> ( 'a1 -> 'a2 t ) -> 'a2 t

val (>>) : 'a1 t -> 'a2 t -> 'a2 t

val (|>>) : 'a t -> ( 'a -> 'b ) -> 'b t

val foldM : ( 'a2 -> 'a1 -> 'a2 t ) -> 'a2 -> 'a1 list -> 'a2 t

val fold1M : ( 'a -> 'a -> 'a t ) -> 'a list -> 'a t

val mapM : ( 'a -> 'b t ) -> 'a list -> 'b list t

val and* : 'a1 t -> 'a2 t -> ('a1 * 'a2) t

val and+ : 'a1 t -> 'a2 t -> ('a1 * 'a2) t